Move HeinleinDataObj out of catalog.py
